Refactoring monitoring code to new class
